pyproject.toml
[project]
name = "sql-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.60.0",
    "fastapi>=0.116.1",
    "fastapi-mcp-client>=0.4.0",
    "httpx>=0.28.1",
    "langchain>=0.2.17",
    "langchain-community>=0.2.19",
    "langchain-core>=0.3.72",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.1.3",
    "langchain-openai>=0.3.28",
    "langgraph>=0.3.25",
    "mcp-client>=0.0.0",
    "mcp[cli]>=1.12.3",
    "oracledb>=3.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
